+++ b3j0f/schema/utils.py
@@ -80,6 +80,7 @@ def datatype2schemacls(
         result = gscls(_datatype, besteffort=_besteffort)
 
     if result is None and _force:
+        print('OK')
         _build = build if _factory is None else _factory.build
 
         result = _build(_resource=_datatype, **kwargs)
@@ -330,7 +331,7 @@ def updatecontent(schemacls=None, updateparents=True, exclude=None):
             schemacls=schemacls, updateparents=updateparents, exclude=exclude
         )
 
-    if updateparents:
+    if updateparents and hasattr(schemacls, 'mro'):
         schemaclasses = reversed(list(schemacls.mro()))
 
     else: