+++ changelog.rst
@@ -1,6 +1,12 @@
 ChangeLog
 =========
 
+0.0.4 (2016/08/08)
+------------------
+
+- fix schema class without mro (python2).
+- raise an exception while trying to import a type defined in a function signature not present in the python execution scope.
+
 0.0.3 (2016/08/08)
 ------------------